Create UnityContainerHolder
TK: 420507
