+++ Server/Server.Console/Program.cs
@@ -5,12 +5,8 @@
 // <author>Myroslava Tarcha</author>
 
 using System;
-using System.Configuration;
 using System.Data.Entity;
-using System.Diagnostics.CodeAnalysis;
 using Common.DatabaseRepositories;
-using Microsoft.Practices.Unity;
-using Microsoft.Practices.Unity.Configuration;
 using NLog;
 
 namespace Server.Console
@@ -20,8 +16,6 @@ namespace Server.Console
     /// </summary>
     public static class Program
     {
-        private const string UnityContainerFile = "Unity.config";
-
         private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
 
         /// <summary>
@@ -33,28 +27,7 @@ namespace Server.Console
 
             AppDomain.CurrentDomain.UnhandledException += OnUnhandledException;
 
-            using (var container = CreateUnityContainer())
-            {
-            }
-
             LogManager.Flush();
-        }
-
-        [SuppressMessage("Microsoft.Reliability", "CA2000:Dispose objects before losing scope", Justification = "The container must be disposed in external code.")]
-        private static UnityContainer CreateUnityContainer()
-        {
-            var fileMap = new ExeConfigurationFileMap
-            {
-                ExeConfigFilename = UnityContainerFile
-            };
-
-            var configuration = ConfigurationManager.OpenMappedExeConfiguration(fileMap, ConfigurationUserLevel.None);
-            var unitySection = (UnityConfigurationSection)configuration.GetSection("unity");
-
-            var container = new UnityContainer();
-            container.LoadConfiguration(unitySection);
-
-            return container;
         }
 
         private static void OnUnhandledException(object sender, UnhandledExceptionEventArgs e)
