--- Server/Server.Console/Unity.config
@@ -1,11 +1,0 @@
-﻿<configuration>
-  <configSections>
-    <section name="unity" type="Microsoft.Practices.Unity.Configuration.UnityConfigurationSection, Microsoft.Practices.Unity.Configuration"/>
-  </configSections>
-  <unity xmlns="http://schemas.microsoft.com/practices/2010/unity">
-
-    <container>
-    </container>
-
-  </unity>
-</configuration>
